refactor: Gacha 타입 분리
